moving all unit tests to unittests.py
